requirements.txt
Django==4.2.9
PyMySQL==1.1.0
Pillow==10.1.0
django-cors-headers==4.3.1
